Refactored out xlsx specific functions into separate module.
